minor adjustments to hash table sizing

+++ src/kcount-gpu/gpu_hash_table.hpp
@@ -120,7 +120,7 @@ class HashTableGPUDriver {
   HashTableGPUDriver();
   ~HashTableGPUDriver();
 
-  bool init(int upcxx_rank_me, int upcxx_rank_n, int kmer_len, int max_elems, int gpu_avail_mem, double &init_time,
+  bool init(int upcxx_rank_me, int upcxx_rank_n, int kmer_len, int max_elems, size_t gpu_avail_mem, double &init_time,
             size_t &gpu_bytes_reqd);
 
   void insert_kmer(const uint64_t *kmer, uint16_t kmer_count, char left, char right);
